Correct behaviour near leap seconds, gaps and ambiguous times More tests for TAI
git-svn-id: https://threeten.svn.sourceforge.net/svnroot/threeten/trunk/jsr-310-ri@907 291d795c-afe8-5c46-8ee5-bf9dd72e1864

+++ src/main/java/javax/time/scales/LeapSeconds.java
@@ -75,7 +75,7 @@ public class LeapSeconds {
         @Override
         void initialise(UTCPeriod previous) {
             deltaSeconds = previous == null ? 10 : ((Entry)previous).deltaSeconds+1;
-            deltaNanoseconds = deltaSeconds * ScaleUtil.NANOS_PER_SECOND;
+            deltaNanoseconds = deltaSeconds * (long)ScaleUtil.NANOS_PER_SECOND;
             super.initialise(previous);
         }
